refactor(training): use step decorator for training step orchestration & define steps in the trainer class
