+++ application/backend/app/core/run/runnable.py
@@ -14,7 +14,6 @@ Classes:
     RunnableFactory: Factory for registering and creating Runnable instances by type.
 """
 
-import logging
 from collections.abc import Callable
 from dataclasses import dataclass
 from typing import Generic, Protocol, TypeVar
@@ -22,19 +21,12 @@ from typing import Generic, Protocol, TypeVar
 ReportFn = Callable[[str, float], None]
 HeartbeatFn = Callable[[], None]
 
-logger = logging.getLogger(__name__)
-
 
 @dataclass(kw_only=True)
 class ExecutionContext:
     payload: str
     report: ReportFn
     heartbeat: HeartbeatFn
-
-    def report_progress(self, msg: str = "", progress: float = 0.0) -> None:
-        """Report progress of the execution."""
-        logger.info(msg)
-        self.report(msg, progress)
 
 
 class Runnable(Protocol):  # ignore
