+++ application/backend/app/services/training/__init__.py
@@ -2,6 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .base import Trainer
-from .otx import OTXTrainer
+from .otx_trainer import OTXTrainer
 
 __all__ = ["OTXTrainer", "Trainer"]
