--- application/backend/app/services/training/steps/__init__.py
@@ -1,12 +1,0 @@
-# Copyright (C) 2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .assign_subsets import AssignSubsetsStep
-from .otx_train_model import OTXTrainModelStep
-from .prepare_weights import PrepareWeightsStep
-
-__all__ = [
-    "AssignSubsetsStep",
-    "OTXTrainModelStep",
-    "PrepareWeightsStep",
-]
