+++ application/backend/tests/integration/services/training/steps/subset_assignment/test_subset_service.py
@@ -9,8 +9,7 @@ from sqlalchemy.orm import Session
 
 from app.db.schema import DatasetItemDB, LabelDB, ProjectDB
 from app.models import DatasetItemSubset
-from app.services.training.steps.subset_assignment import SubsetService
-from app.services.training.steps.subset_assignment.models import SubsetAssignment
+from app.services.training.subset_assignment import SubsetAssignment, SubsetService
 from tests.integration.project_factory import ProjectTestDataFactory
