+++ application/backend/tests/unit/services/training/subset_assignment/test_distribution.py
@@ -4,7 +4,7 @@
 import pytest
 
 from app.models import DatasetItemSubset
-from app.services.training.steps.subset_assignment import SplitRatios, SubsetDistribution
+from app.services.training.subset_assignment import SplitRatios, SubsetDistribution
 
 
 class TestSubsetDistribution:
